+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+-   Assign IPs to subnets with command "fixip"
 -   Display error when user asks for specific category entries but category is not assigned to object type(s)